requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.15
python-dotenv>=1.0.0
deep-translator>=1.11.4
openai>=1.12.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-search-results>=2.4.2
serpapi>=0.1.0
tavily-python>=0.3.1
pytz>=2023.3
python-dateutil>=2.8.2
sentence-transformers
torch